Add connector documentation links
